Convert script_apply_edits tool
